add retry for tx sending
